gateway: emit integer Responses timestamps across HTTP and WebSocket (#781)
* gateway: emit integer created_at and completed_at on Responses envelopes

OpenAI emits both fields as integer epoch seconds. Strict typed clients
(Goose's Rust Responses decoder) reject the float the gateway used to
send, which made every gpt- alias unusable through Goose. The Python
and Rust Responses encoders now take integer seconds (floor); the
Chat Completions integer created is unchanged. Parity goldens and
fixtures move from fractional to whole seconds.


* gateway: move remaining Responses fixtures to integer seconds


* gateway: assert integer prewarm ack timestamps on the Responses WebSocket


* gateway: bump native crate to 0.3.30 for the Responses timestamp change


* gateway: use integer created_at in the Rust Responses encoder unit tests


* gateway: integer created_at in the Rust Responses tests merged from main


* gateway: use integer timestamps in newly merged Responses fixtures


* Keep uv.lock aligned with main apart from version bumps


* gateway: use integer timestamps in newly merged Rust encoder tests


---------

+++ exp/runtime/gateway/native/src/lib.rs
@@ -260,7 +260,7 @@ fn metrics_snapshot_json() -> String {
 fn encode_responses_fixture(
     request_id: &str,
     model: &str,
-    created_at: f64,
+    created_at: i64,
     envelope_json: &str,
     events_json: &str,
 ) -> PyResult<Vec<String>> {
@@ -288,7 +288,7 @@ fn encode_responses_fixture(
 fn completed_responses_fixture(
     request_id: &str,
     model: &str,
-    created_at: f64,
+    created_at: i64,
     envelope_json: &str,
     events_json: &str,
 ) -> PyResult<String> {
